+++ R/isNodeAlive.R
@@ -25,7 +25,7 @@
 #' \donttest{
 #' cl <- makeClusterPSOCK(2)
 #' 
-#' ## Check if cluster nodes #2 is alive
+#' ## Check if cluster node #2 is alive
 #' print(isNodeAlive(cl[[2]]))
 #' 
 #' ## Check all nodes