+++ NEWS.md
@@ -11,7 +11,7 @@
    typically SSH, and check if the process ID of the worker still
    exists.
 
- * Add `cloneNode()`, which can be used to "restart" a `RickhSOCKnode`
+ * Add `cloneNode()`, which can be used to "restart" `RickhSOCKnode`
    cluster nodes.
 
  * Argument `worker` for `makeNodePSOCK()` now takes the optional,